refactor : 로또 도메인 분리
