+++ src/main/java/lotto/view/OutputView.java
@@ -1,6 +1,7 @@
 package lotto.view;
 
 import java.util.List;
+import lotto.domain.Lottos;
 
 public class OutputView {
 
@@ -8,10 +9,7 @@ public class OutputView {
         System.out.println(String.format("%d개를 구매했습니다.", count));
     }
 
-    public void printLottos(List<List<Integer>> lottos) {
-        for (int i = 0; i < lottos.size(); i ++) {
-            System.out.println(lottos.get(i));
-        }
-        System.out.println();
+    public void printLottos(Lottos lottos) {
+        System.out.println(lottos.toString());
     }
 }
